qaqc/requirements.txt
# QAQC Dependencies
# Quality Assurance & Quality Control scripts requirements

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Geospatial libraries
geopandas>=0.13.0
rasterio>=1.3.0
shapely>=2.0.0

# Visualization
matplotlib>=3.7.0
pillow>=10.0.0

# File system and utilities
pathlib2>=2.3.0  # For Python < 3.4 compatibility if needed
tqdm>=4.65.0    # Progress bars

# Optional: For advanced analysis
scipy>=1.11.0
scikit-image>=0.21.0</content>
<parameter name="filePath">/home/bny/Github/greenland-glacier-flow/qaqc/requirements.txt